move prometheus read endpoint to /prometheus/read
