+++ webapp/graphite/prometheus/urls.py
@@ -1,0 +1,6 @@
+from django.conf.urls import url
+from . import views
+
+urlpatterns = [
+    url(r'^/read/?$', views.renderPrometheus, name='render_prometheus'),
+]
