+++ webapp/graphite/urls.py
@@ -30,6 +30,7 @@ graphite_urls = [
     url('^events', include('graphite.events.urls')),
     url('^tags', include('graphite.tags.urls')),
     url('^functions', include('graphite.functions.urls')),
+    url('^prometheus', include('graphite.prometheus.urls')),
     url('^s/(?P<path>.*)', shorten, name='shorten'),
     url('^S/(?P<link_id>[a-zA-Z0-9]+)/?$', follow, name='follow'),
     url('^$', browser, name='browser'),
